+++ Changes.md
@@ -1,7 +1,10 @@
 1.6.x.x (relative to 1.6.1.0)
 =======
 
+Fixes
+-----
 
+- SceneInspector : Fixed units used to show transform rotation - it is now shown in degrees again (#6604).
 
 1.6.1.0 (relative to 1.6.0.0)
 =======