+++ Changes.md
@@ -11,6 +11,7 @@ Fixes
 -----
 
 - SceneInspector : Fixed units used to show transform rotation - it is now shown in degrees again (#6604).
+- HierarchyView, LightEditor, AttributeEditor : Fixed context used by the visibility column to evaluate the scene.
 - OSLShader : Fixed crash loading string array parameters without default values for all elements.
 
 1.6.1.0 (relative to 1.6.0.0)